Set customOpenshiftSelect value as initial (when applicable)
